refactor: LottoGame 클래스 controller 패키지로 이동 후 클래스 명 변경 (LottoGame -> LottoController)
